+++ src/modules/auth/dtos/auth-login.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsEmail } from 'class-validator';
+import { IsEmail, IsString } from 'class-validator';
 
 export class AuthLoginDto {
   @ApiProperty({
@@ -13,5 +13,6 @@ export class AuthLoginDto {
     type: 'string',
     example: '12345678',
   })
+  @IsString()
   password: string;
 }
