+++ src/modules/auth/strategies/local.strategy.ts
@@ -1,8 +1,8 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
-import { AuthService } from './auth.service';
-import { UserDto } from './dtos/user.dto';
+import { AuthService } from '../auth.service';
+import { UserDto } from '../dtos/user.dto';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
