+++ src/modules/auth/strategies/optional-jwt.strategy.ts
@@ -1,14 +1,22 @@
-import { Injectable, Inject, UnauthorizedException } from '@nestjs/common';
+import {
+  Injectable,
+  Inject,
+  InternalServerErrorException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { ConfigType } from '@nestjs/config';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt } from 'passport-jwt';
 import { Strategy } from 'passport-jwt';
 import { jwtConfig } from 'src/config/jwt.config';
-import { AuthService } from './auth.service';
-import { JwtPayload } from './interfaces/jwt-payload.interface';
+import { AuthService } from '../auth.service';
+import { JwtPayload } from '../interfaces/jwt-payload.interface';
 
 @Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy) {
+export class OptionalJwtStrategy extends PassportStrategy(
+  Strategy,
+  'optional-jwt',
+) {
   constructor(
     private authService: AuthService,
     @Inject(jwtConfig.KEY)
@@ -22,10 +30,14 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
 
   async validate(payload: JwtPayload) {
     try {
+      if (!payload) {
+        return null;
+      }
+
       const user = await this.authService.findById(payload.sub);
 
       if (!user) {
-        throw new UnauthorizedException();
+        return null;
       }
 
       return user;
